Add support for Sapi-NT hierarchy endpoints
This commit enables sapi-client to support Sapi-NT hierarchy endpoints.
These endpoints present hierarchically-organised resources, such as
terms in a SKOS hierarchy. We want to be able to support such
hierarchies by:

- getting all terms in the hierarchy at once
- finding the roots of the hierarchy
- for any node in the hierarchy, knowing its parent and children
- being able to travserse the hierarchy breadth-first or depth-first
  to access the hierarchy elements in some order
- support SKOS hierarchies as a default
- allowing for other hierarchy encoding schemes.

We add a sub-class of `SapiClient::Resource` -
`SapiClient::HierarchyResource` which will encapsulate resources that
are contained in a hierarchy, and add a `{ep}_hierarchy` method to
instances for some endpoint name `ep` to access the hierarchy directly.