Allow snake_case method names for properties
To address issue #21, this commit allows a method name to a
`SapiResource` object to be in `snake_case`, even if the corresponding
property name is in camelCase (which is normal for RDF properties).

For example if this resource is wrapped as a `SapiResource`:

```json
{
    "prefLabel": "Hello world"
}
```

it can be accessed via:

```ruby
r = SapiResource.new(...)
r.pref_label

=> 'Hello world'
```

Prior to this commit, the method name would have needed to be
`prefLabel`, which provokes complaints from the Rubocop linter.