Don't log API calls by default in Rails production
Ticket #26

If we are using Sapi-client in the context of a Rails application, we
can make an assumption that our current logging practices are sufficient
and therefore we don't need to log the interaction with the API. Since
this sometimes may not be true, it can be overridden by setting the
`Rails.application.config.sapi_client_log_api_calls` config variable to
`true`.

Most of the code in this PR is in the tests, as we need to set up a
global `Rails` instance that the logger can interrogate, but ensure that
global state is cleaned up at the end of the test.

I also took the decision to perform the test by inspecting the return
result from the private method `rails_logging?`. Although we don't
normally test private methods, in this case it made for a more compact
(!) and more meaningful test, rather than attempting to actually perform
an API call and inspecting the log output. This feels more like testing
Faraday than this library.

A possible refactoring cleanup in future is to extract the
`rails_logging?` logic into a class with a public API, and then have a
private instance of that class within the `Instance` object.