Add a single global register of endpoint views
To address #44, we add a single shared global instance that records
views by name. Previously, endpoint groups kept a Hash of the views
defined in them, allowing a certain degree of cross-referencing.
However, in SapiNT the view namespace is global, so one endpoint can
refer to a view defined in an entirely separate endpoint file.

To address this, we remove the per-endpoint Hash of views, and add a
single shared global instance of a `ViewRegistry`.

In order to test the specific regression, this commit also adds the
SapiNT spec for CBD as a fixture, since this is where #44 was first
reported.